+++ CHANGELOG.md
@@ -1,3 +1,22 @@
-## 1.0.0
+# Changelog
 
-- Initial release
+*Changelog created using the [Simple Changelog](https://marketplace.visualstudio.com/items?itemName=tobiaswaelde.vscode-simple-changelog) extension for VS Code.*
+
+## [1.4] - 2025-01-09
+### Added
+- Add runMock command
+
+### Changed
+- Split extension.js file
+
+
+## [1.3] - 2024-12-27
+
+
+## [1.2]
+
+
+## [1.1]
+
+
+## [1.0]